requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
PyYAML==6.0.1
structlog==23.2.0
python-jose[cryptography]==3.3.0
requests-oauthlib==1.3.1
atlassian-python-api==3.41.10